Add update from local sd card

+++ src/www/httpd/cgi-bin/fw_upgrade.sh
@@ -25,13 +25,20 @@ fi
 if [ "$VAL" == "info" ] ; then
     printf "Content-type: application/json\r\n\r\n"
 
-    FW_VERSION=`cat /home/yi-hack/version`
+    MODEL_SUFFIX=`cat $YI_HACK_PREFIX/model_suffix`
+    FW_VERSION=`cat $YI_HACK_PREFIX/version`
     LATEST_FW=`/home/yi-hack/usr/bin/wget -O -  https://api.github.com/repos/roleoroleo/yi-hack-Allwinner/releases/latest 2>&1 | grep '"tag_name":' | sed -E 's/.*"([^"]+)".*/\1/'`
+    if [ -f /tmp/sd/${MODEL_SUFFIX}_x.x.x.tgz ]; then
+        LOCAL_FW="true"
+    else
+        LOCAL_FW="false"
+    fi
 
     printf "{\n"
     printf "\"%s\":\"%s\",\n" "error" "false"
     printf "\"%s\":\"%s\",\n" "fw_version"      "$FW_VERSION"
-    printf "\"%s\":\"%s\"\n" "latest_fw"       "$LATEST_FW"
+    printf "\"%s\":\"%s\",\n" "latest_fw"       "$LATEST_FW"
+    printf "\"%s\":%s\n" "local_fw"             "$LOCAL_FW"
     printf "}"
 
 elif [ "$VAL" == "upgrade" ] ; then
@@ -58,19 +65,24 @@ elif [ "$VAL" == "upgrade" ] ; then
     cd /tmp/sd/fw_upgrade
 
     MODEL_SUFFIX=`cat $YI_HACK_PREFIX/model_suffix`
-    FW_VERSION=`cat /home/yi-hack/version`
-    LATEST_FW=`/home/yi-hack/usr/bin/wget -O -  https://api.github.com/repos/roleoroleo/yi-hack-Allwinner/releases/latest 2>&1 | grep '"tag_name":' | sed -E 's/.*"([^"]+)".*/\1/'`
-    if [ "$FW_VERSION" == "$LATEST_FW" ]; then
-        printf "Content-type: text/html\r\n\r\n"
-        printf "No new firmware available."
-        exit
-    fi
+    FW_VERSION=`cat $YI_HACK_PREFIX/version`
+    if [ -f /tmp/sd/${MODEL_SUFFIX}_x.x.x.tgz ]; then
+        mv /tmp/sd/${MODEL_SUFFIX}_x.x.x.tgz /tmp/sd/fw_upgrade/${MODEL_SUFFIX}_x.x.x.tgz
+        LATEST_FW="x.x.x"
+    else
+        LATEST_FW=`/home/yi-hack/usr/bin/wget -O -  https://api.github.com/repos/roleoroleo/yi-hack-Allwinner/releases/latest 2>&1 | grep '"tag_name":' | sed -E 's/.*"([^"]+)".*/\1/'`
+        if [ "$FW_VERSION" == "$LATEST_FW" ]; then
+            printf "Content-type: text/html\r\n\r\n"
+            printf "No new firmware available."
+            exit
+        fi
 
-    /home/yi-hack/usr/bin/wget https://github.com/roleoroleo/yi-hack-Allwinner/releases/download/$LATEST_FW/${MODEL_SUFFIX}_${LATEST_FW}.tgz
-    if [ ! -f ${MODEL_SUFFIX}_${LATEST_FW}.tgz ]; then
-        printf "Content-type: text/html\r\n\r\n"
-        printf "Unable to download firmware file."
-        exit
+        /home/yi-hack/usr/bin/wget https://github.com/roleoroleo/yi-hack-Allwinner/releases/download/$LATEST_FW/${MODEL_SUFFIX}_${LATEST_FW}.tgz
+        if [ ! -f ${MODEL_SUFFIX}_${LATEST_FW}.tgz ]; then
+            printf "Content-type: text/html\r\n\r\n"
+            printf "Unable to download firmware file."
+            exit
+        fi
     fi
 
     tar zxvf ${MODEL_SUFFIX}_${LATEST_FW}.tgz
